denigma/requirements/project.txt
# project.txt is a pip requirements file which describes the distributions
# required by your project to run.

--requirement=base.txt

# Put project-specific requirements here.
# See http://pip-installer.org/requirement-format.html for more information.

markdown	# Markup for Wiki.
south 		# Necessary for database evolution.
MySQL-python 	# Required for interfacing with db via manage.py on server.
PIL		# Used by gallery.

-e git+https://github.com/boto/boto.git#egg=boto # AWS API. 
django-SES      # Simple Email Service.
django-reversion==1.6.3 # Model version control system. #==1.5.3 # Worked for Django-1.3.2 #==1.6.3 works for Django-1.4.1
docutils # Documentation utilities.

#scipy # Required for parts of the stats app. However, scipy can cause issues on deployment (see docs).
biopython # For EUtils
cogent # For t-test
intermine # Provides yeastmine (SGD) API.

# Tags:
django-taggit   # Provides tagging functionality.
-e svn+http://django-tagging.googlecode.com/svn/trunk/#egg=tagging # Allows the inclusion of tags to any model, is used for links.
pip install -e git+https://github.com/brosner/django-tagging#egg=tagging


# Alternative Admin designs:
django-grappelli==2.3.8 # Alternative admin design with additional functionality.
#-e git://github.com/divio/djangocms-admin-style.git#egg=djangocms_admin_style
#-e git://github.com/riccardo-forina/django-admin-bootstrapped#egg=django-admin-bootstrapped

# Searching:
whoosh
-e git://github.com/toastdriven/django-haystack.git@master#egg=django-haystack

# Web crawling:
BeautifulSoup
bottlenose # Wrapper for Amazon product advertising API.
python-amazon-product-api # For retrieving information on Books from Amazon.
mechanize

# Tables:
#django-tables2
-e git://github.com/bradleyayers/django-tables2.git#egg=django-tables2

# Forms:
django-crispy-forms #==dev # Dev does not work.

# Filters:
-e git://github.com/alex/django-filter.git#egg=django-filters # Not django-1.3 compatibile
#-e git://github.com/hevok/django-filter.git#egg=django-filters
django_easyfilters
#django-ajax-filtered-fields

# Hierarchy:
-e git+https://github.com/django-mptt/django-mptt/#egg=django-mptt

-e git://github.com/bradleyayers/suds-htj.git#egg=suds # Required for DAVID. suds==0.4 does not work with DjDT

# Real-time previewing of Markup:
-e git://github.com/timmyomahony/django-pagedown.git#egg=pagedown

# PDF generation:
reportlab==2.5

# Source code highlighting:
pygments

# Representative user images:
-e git://github.com/guzru/django-avatar#egg=django-avatar

# Comments:
-e git://github.com/edoburu/django-fluent-comments.git#egg=django-fluent-comments


# Chat
gevent-socketio==0.3.5-beta
irc==2.0
daemon==1.0
sphinx-me
